feat: 영화 추가 기능 구현 및 이미지 url 추가
