--- app/user/models.py
@@ -1,22 +1,0 @@
-from sqlalchemy import Integer, String, ForeignKey
-from sqlalchemy.orm import relationship, Mapped, mapped_column
-from datetime import datetime
-from database.common import Base
-from sqlalchemy import DateTime
-from app.review.models import Review
-
-class User(Base):
-    __tablename__ = 'user'
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    username: Mapped[str] = mapped_column(String(50), nullable=False)
-    login_id: Mapped[str] = mapped_column(String(50), nullable=False)
-    hashed_pwd: Mapped[str] = mapped_column(String(100), nullable=False)
-
-    reviews: Mapped[list["Review"]] = relationship(back_populates="user")
-    
-class BlockedToken(Base):
-    __tablename__ = "blocked_token"
-
-    token_id: Mapped[str] = mapped_column(String(255), primary_key=True)
-    expired_at: Mapped[datetime] = mapped_column(DateTime)
