+++ watchapedia/app/user/models.py
@@ -1,0 +1,19 @@
+from sqlalchemy import Integer, String, ForeignKey
+from sqlalchemy.orm import relationship, Mapped, mapped_column
+from datetime import datetime
+from watchapedia.database.common import Base
+from sqlalchemy import DateTime
+
+class User(Base):
+    __tablename__ = 'user'
+
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    username: Mapped[str] = mapped_column(String(50), nullable=False)
+    login_id: Mapped[str] = mapped_column(String(50), nullable=False)
+    hashed_pwd: Mapped[str] = mapped_column(String(100), nullable=False)
+    
+class BlockedToken(Base):
+    __tablename__ = "blocked_token"
+
+    token_id: Mapped[str] = mapped_column(String(255), primary_key=True)
+    expired_at: Mapped[datetime] = mapped_column(DateTime)
