+++ watchapedia/common/errors.py
