+++ watchapedia/database/__init__.py
@@ -1,0 +1,6 @@
+import watchapedia.app.user.models
+import watchapedia.app.movie.models
+import watchapedia.app.genre.models
+import watchapedia.app.country.models
+import watchapedia.app.participant.models
+import watchapedia.app.review.models
